Remove asyncawait package
asyncawait package use the tokens await and async which are reserved
words in ES6. This makes it difficult to use dbffile in projects
where the dependencies sources are transpiled (like Meteor).

Also, typescript added support, in 2.1, for await/async when compiling
down to ES3/ES5. So asyncawait is not actually needed anymore.

Drop it and use ES6 async/await feature by upgrading typescript.