fix: remove asyncTester helper function
It is not needed. Mocha can get a promise
from the `it` callback and use it to determine
if the test passed or failed.

async functions already return a promise so
no wrapping has to be done to them.

Also, both returning a promise and calling done
is not the right thing to do.
See: https://github.com/mochajs/mocha/issues/1283#issuecomment-290269883